+++ core/base_backup_test.go
@@ -127,6 +127,7 @@ func verifyBackupContent(app core.App, path string) error {
 		"logs.db-shm",
 		"logs.db-wal",
 		".gitignore",
+		".pb_temp_to_delete",
 	}
 
 	entries, err := os.ReadDir(dir)